requirements.txt
apache-airflow-providers-amazon
confluent-kafka
pandas